+++ src-test/net/soliddesign/iumpr/controllers/MonitorCompletionControllerTest.java
@@ -236,7 +236,7 @@ public class MonitorCompletionControllerTest {
         when(engineSpeedModule.isEngineCommunicating()).thenReturn(true);
         when(comparisonModule.compareFileToVehicle(any(ResultsListener.class), eq(reportFileModule), eq(2), eq(17)))
                 .thenReturn(true);
-        when(reportFileModule.getMinutesSinceCodeClear()).thenReturn(1234);
+        when(reportFileModule.getMinutesSinceCodeClear()).thenReturn((double) 1234);
 
         LocalDateTime monitorTime = LocalDateTime.now();
         when(reportFileModule.getInitialMonitorsTime()).thenReturn(monitorTime);
@@ -311,7 +311,7 @@ public class MonitorCompletionControllerTest {
         inOrder.verify(listener).onProgress(10, 17, "Requesting DM21");
         inOrder.verify(reportFileModule).onProgress(10, 17, "Requesting DM21");
         inOrder.verify(reportFileModule).getMinutesSinceCodeClear();
-        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq(1234));
+        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq((double) 1234));
 
         inOrder.verify(listener).onProgress(11, 17, "Tracking Monitor Completion Status");
         inOrder.verify(reportFileModule).onProgress(11, 17, "Tracking Monitor Completion Status");
@@ -350,7 +350,7 @@ public class MonitorCompletionControllerTest {
         inOrder.verify(reportFileModule).onResult("");
         inOrder.verify(listener).onProgress(14, 17, "Requesting DM21");
         inOrder.verify(reportFileModule).onProgress(14, 17, "Requesting DM21");
-        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq(-1));
+        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq((double) -1));
 
         inOrder.verify(listener).onResult("");
         inOrder.verify(reportFileModule).onResult("");
@@ -389,7 +389,7 @@ public class MonitorCompletionControllerTest {
         when(engineSpeedModule.isEngineCommunicating()).thenReturn(true);
         when(comparisonModule.compareFileToVehicle(any(ResultsListener.class), eq(reportFileModule), eq(2), eq(17)))
                 .thenReturn(true);
-        when(reportFileModule.getMinutesSinceCodeClear()).thenReturn(1234);
+        when(reportFileModule.getMinutesSinceCodeClear()).thenReturn((double) 1234);
         when(dateTimeModule.getDateTime()).thenReturn("Time").thenReturn("Time");
         when(monitorTrackingModule.getLastDm5Time()).thenReturn(null);
 
@@ -443,7 +443,7 @@ public class MonitorCompletionControllerTest {
         inOrder.verify(listener).onProgress(10, 17, "Requesting DM21");
         inOrder.verify(reportFileModule).onProgress(10, 17, "Requesting DM21");
         inOrder.verify(reportFileModule).getMinutesSinceCodeClear();
-        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq(1234));
+        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq((double) 1234));
 
         inOrder.verify(listener).onProgress(11, 17, "Tracking Monitor Completion Status");
         inOrder.verify(reportFileModule).onProgress(11, 17, "Tracking Monitor Completion Status");
@@ -471,7 +471,7 @@ public class MonitorCompletionControllerTest {
         inOrder.verify(reportFileModule).onResult("");
         inOrder.verify(listener).onProgress(14, 17, "Requesting DM21");
         inOrder.verify(reportFileModule).onProgress(14, 17, "Requesting DM21");
-        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq(-1));
+        inOrder.verify(diagnosticReadinessModule).reportDM21(any(ResultsListener.class), eq((double) -1));
 
         inOrder.verify(listener).onResult("");
         inOrder.verify(reportFileModule).onResult("");
